Silence compiler warnings in -Wformat mode.
Passing SQLINTEGER/SQLUINTEGER sprintf "%d", "%ld" modifiers gave a warning
on some systems (at least on my 32-bit Debian system) .

+++ convert.c
@@ -4219,27 +4219,27 @@ mylog("C_WCHAR=%s(%d)\n", buffer, used);
 		}
 		case SQL_C_INTERVAL_YEAR:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d years", ivsign, ivstruct->intval.year_month.year);
+			sprintf(param_string, "%s%u years", ivsign, (unsigned int) ivstruct->intval.year_month.year);
 			break;
 		case SQL_C_INTERVAL_MONTH:
 		case SQL_C_INTERVAL_YEAR_TO_MONTH:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d years %s%d mons", ivsign, ivstruct->intval.year_month.year, ivsign, ivstruct->intval.year_month.month);
+			sprintf(param_string, "%s%u years %s%u mons", ivsign, (unsigned int) ivstruct->intval.year_month.year, ivsign, (unsigned int) ivstruct->intval.year_month.month);
 			break;
 		case SQL_C_INTERVAL_DAY:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d days", ivsign, ivstruct->intval.day_second.day);
+			sprintf(param_string, "%s%u days", ivsign, (unsigned int) ivstruct->intval.day_second.day);
 			break;
 		case SQL_C_INTERVAL_HOUR:
 		case SQL_C_INTERVAL_DAY_TO_HOUR:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d days %s%02d:00:00", ivsign, ivstruct->intval.day_second.day, ivsign, ivstruct->intval.day_second.hour);
+			sprintf(param_string, "%s%u days %s%02u:00:00", ivsign, (unsigned int) ivstruct->intval.day_second.day, ivsign, (unsigned int) ivstruct->intval.day_second.hour);
 			break;
 		case SQL_C_INTERVAL_MINUTE:
 		case SQL_C_INTERVAL_HOUR_TO_MINUTE:
 		case SQL_C_INTERVAL_DAY_TO_MINUTE:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d days %s%02d:%02d:00", ivsign, ivstruct->intval.day_second.day, ivsign, ivstruct->intval.day_second.hour, ivstruct->intval.day_second.minute);
+			sprintf(param_string, "%s%u days %s%02u:%02u:00", ivsign, (unsigned int) ivstruct->intval.day_second.day, ivsign, (unsigned int) ivstruct->intval.day_second.hour, (unsigned int) ivstruct->intval.day_second.minute);
 			break;
 
 		case SQL_C_INTERVAL_SECOND:
@@ -4247,11 +4247,11 @@ mylog("C_WCHAR=%s(%d)\n", buffer, used);
 		case SQL_C_INTERVAL_HOUR_TO_SECOND:
 		case SQL_C_INTERVAL_MINUTE_TO_SECOND:
 			ivsign = ivstruct->interval_sign ? "-" : "";
-			sprintf(param_string, "%s%d days %s%02d:%02d:%02d",
-				ivsign, ivstruct->intval.day_second.day,
-				ivsign, ivstruct->intval.day_second.hour,
-				ivstruct->intval.day_second.minute,
-				ivstruct->intval.day_second.second);
+			sprintf(param_string, "%s%u days %s%02u:%02u:%02u",
+					ivsign, (unsigned int) ivstruct->intval.day_second.day,
+					ivsign, (unsigned int) ivstruct->intval.day_second.hour,
+					(unsigned int) ivstruct->intval.day_second.minute,
+					(unsigned int) ivstruct->intval.day_second.second);
 			if (ivstruct->intval.day_second.fraction > 0)
 			{
 				int fraction = ivstruct->intval.day_second.fraction, prec = apara->precision;
